rabbit clients optimization
