+++ gateway/payment_gate.go
@@ -38,7 +38,7 @@ func (gw *Gateway) PaymentGateway(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	if err := gw.mq.RunPaymentCheckoutTasks(orderId, accessToken, payload.ProductId); err != nil {
+	if err := gw.RunPaymentCheckoutTasks(orderId, accessToken, payload.ProductId); err != nil {
 		utils.WriteJSON(w, map[string]string{
 			"message":    "Payment could not be processed",
 			"error_info": err.Error(),
